requirements.txt
yfinance==0.1.74
mplcyberpunk==0.7.1
python-dotenv==1.0.1